bugFixInDeletionCommit

+++ app.js
@@ -195,16 +195,39 @@ app.post("/", function(req, res){
 
 app.post("/delete", function(req, res){
   const checkedItemId = req.body.checkbox;
+  const listName = req.body.listName;
+
+  if (listName === "Today") {
+    Item.findByIdAndRemove(checkedItemId)
+      .then(() => {
+        console.log("Successfully deleted checked item.");
+        res.redirect("/");
+      })
+      .catch((err) => {
+        console.log(err);
+      });
+  } else {
+    List.findOneAndUpdate({ name: listName }, { $pull: { items: { _id: checkedItemId } } })
+    .then(foundList => {
+      res.redirect("/" + listName);
+    })
+    .catch(err => {
+      console.log(err);
+    });
+
+
+
+
+
+    // List.findOneAndUpdate({name: listName}, {$pull: {items: {_id: checkedItemId}}}, function(err, foundList) {
+    //   if (!err) {
+    //     res.redirect("/" + listName);
+    //   }
+    // });
+  }
   // Item.findByIdAndRemove(checkedItemId, console.log("Succesfully deleted checked item."));
   // res.redirect("/")
-  Item.findByIdAndRemove(checkedItemId)
-  .then(() => {
-    console.log("Successfully deleted checked item.");
-    res.redirect("/");
-  })
-  .catch((err) => {
-    console.log(err);
-  });
+
 
 });
